9C-1039-1 Repair errors apply code review
Some errors were due to a transitive dependencies to Guava, a Google
library, for which an older version was being used. We cut that.

We apply code review comments.
We use command-line based generator for private keys.
