+++ modules/commons/src/test/scala/cards/nine/domain/ScalaCheck.scala
@@ -3,11 +3,13 @@ package cards.nine.domain
 import cards.nine.domain.account.Email
 import cards.nine.domain.analytics.DateRange
 import cards.nine.domain.application.{ Category, Moment, Package, PriceFilter }
+import cards.nine.domain.oauth.ServiceAccount
 import com.fortysevendeg.scalacheck.datetime.instances.joda._
 import com.fortysevendeg.scalacheck.datetime.GenDateTime.genDateTimeWithinRange
 import enumeratum.{ Enum, EnumEntry }
 import org.joda.time.{ DateTime, Period }
 import org.scalacheck.{ Arbitrary, Gen }
+import sys.process._
 
 object ScalaCheck {
 
@@ -61,6 +63,46 @@ object ScalaCheck {
     (10, Gen.numChar),
     (6, choose(97.toChar, 102.toChar))
   )
+
+  val genServiceAccount: Gen[ServiceAccount] = {
+    // We get these formats from reading one example of credentials
+
+    /*
+     * Generating a PKCS8 Private Key in Base64 format is a lot of work, not worth it
+     * http://stackoverflow.com/questions/5127379/how-to-generate-a-rsa-keypair-with-a-privatekey-encrypted-with-password/6164414#6164414
+     */
+    val privateKeyGen: Gen[String] = {
+      def commandLinePkcs8(i: Int): String = {
+        // This is an ugly hacks but seems to do the job
+        val logger = ProcessLogger(_ ⇒ (), _ ⇒ ()) // ignore all outputs
+        ("openssl genrsa" #| "openssl pkcs8 -topk8 -nocrypt").!!(logger)
+      }
+      Gen.const(1).map(commandLinePkcs8)
+    }
+
+    val privateKeyIdGen: Gen[String] = Gen.listOfN(41, hexChar).map(_.mkString)
+
+    val clientIdGen: Gen[String] = Gen.listOfN(22, Gen.numChar).map(_.mkString)
+
+    for {
+      clientId ← clientIdGen
+      clientEmail ← emailGenerator.map(_.value)
+      privateKey ← privateKeyGen
+      privateKeyId ← privateKeyIdGen
+      tokenUri ← Gen.alphaStr
+      scopes ← Gen.listOf(Gen.alphaStr)
+    } yield ServiceAccount(
+      clientId     = clientId,
+      clientEmail  = clientEmail,
+      privateKey   = privateKey,
+      privateKeyId = privateKeyId,
+      tokenUri     = tokenUri,
+      scopes       = scopes
+    )
+
+  }
+
+  implicit val arbServiceAccount: Arbitrary[ServiceAccount] = Arbitrary(genServiceAccount)
 
 }
