+++ project/Dependencies.scala
@@ -18,7 +18,7 @@ object Dependencies {
   private val cats = "org.typelevel" %% "cats" % Versions.cats
   private val embeddedRedis = "com.orange.redis-embedded" % "embedded-redis" % Versions.embeddedRedis % "test"
   private val flywaydbCore = "org.flywaydb" % "flyway-core" % Versions.flywaydb
-  private val googleApiClient = "com.google.api-client" % "google-api-client" % Versions.googleApiClient
+  private val googleApiClient = "com.google.api-client" % "google-api-client" % Versions.googleApiClient exclude("com.google.guava", "*")
   private val hasher = "com.roundeights" %% "hasher" % Versions.hasher
   private val http4sClient = "org.http4s" %% "http4s-blaze-client" % Versions.http4s
   private val jodaConvert = "org.joda" % "joda-convert" % Versions.jodaConvert
